+++ src/main/java/com/github/shopping_mall_be/repository/ProductRepository.java
@@ -9,12 +9,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     Page<Product> findAllByStockGreaterThan(int stock, Pageable pageable);
-    Optional<Product> findById(Integer productId);
-    // 가격 오름차순
-//    Page<Product> findAllByStockGreaterThanOrderByPriceAsc(int stock, Pageable pageable);
+    Optional<Product> findById(Long productId);
 
-    // 가격 내림차순
-//    Page<Product> findAllByStockGreaterThanOrderByPriceDesc(int stock, Pageable pageable);
 }
